+++ compose.yml
@@ -32,7 +32,6 @@ services:
       # Volume exclusions:
       ## Prevent collisions between host and container node_modules
       - node_modules_ghost_root:/home/ghost/node_modules:delegated
-      - node_modules_ghost_adapter-cache-redis:/home/ghost/ghost/adapter-cache-redis/node_modules:delegated
       - node_modules_ghost_admin:/home/ghost/ghost/admin/node_modules:delegated
       - node_modules_ghost_api-framework:/home/ghost/ghost/api-framework/node_modules:delegated
       - node_modules_ghost_constants:/home/ghost/ghost/constants/node_modules:delegated
@@ -100,9 +99,7 @@ services:
   mysql:
     image: mysql:8.0.35
     container_name: ghost-mysql
-    command: --innodb-buffer-pool-size=1G --innodb-log-buffer-size=500M
-      --innodb-change-buffer-max-size=50 --innodb-flush-log-at-trx_commit=0
-      --innodb-flush-method=O_DIRECT
+    command: --innodb-buffer-pool-size=1G --innodb-log-buffer-size=500M --innodb-change-buffer-max-size=50 --innodb-flush-log-at-trx_commit=0 --innodb-flush-method=O_DIRECT
     ports:
       - 3306:3306
     environment:
@@ -166,7 +163,6 @@ volumes:
   mysql-data: {}
   redis-data: {}
   node_modules_ghost_root: {}
-  node_modules_ghost_adapter-cache-redis: {}
   node_modules_ghost_admin: {}
   node_modules_ghost_api-framework: {}
   node_modules_ghost_constants: {}
