--- ghost/adapter-cache-redis/index.js
@@ -1,1 +1,0 @@
-module.exports = require('./lib/AdapterCacheRedis');
