+++ ghost/core/core/server/adapters/cache/Redis.js
@@ -1,4 +1,4 @@
 /* eslint-disable ghost/filenames/match-regex */
-const RedisCache = require('@tryghost/adapter-cache-redis');
+const RedisCache = require('../lib/redis/AdapterCacheRedis');
 
 module.exports = RedisCache;
