+++ ghost/core/core/server/adapters/lib/redis/AdapterCacheRedis.js
