+++ ghost/core/core/server/adapters/lib/redis/redis-store-factory.js
