+++ ghost/core/package.json
@@ -64,7 +64,6 @@
     "@sentry/node": "7.120.3",
     "@slack/webhook": "7.0.5",
     "@tryghost/adapter-base-cache": "0.1.12",
-    "@tryghost/adapter-cache-redis": "0.0.0",
     "@tryghost/admin-api-schema": "4.5.5",
     "@tryghost/api-framework": "0.0.0",
     "@tryghost/bookshelf-plugins": "0.6.25",
@@ -138,9 +137,12 @@
     "bookshelf-relations": "2.8.0",
     "brute-knex": "4.0.1",
     "bson-objectid": "2.0.4",
+    "cache-manager": "4.1.0",
+    "cache-manager-ioredis": "2.1.0",
     "chalk": "4.1.2",
     "charset": "1.0.1",
     "cheerio": "0.22.0",
+    "cluster-key-slot": "1.1.2",
     "common-tags": "1.8.2",
     "compression": "1.8.0",
     "connect-slashes": "1.4.0",
