+++ ghost/core/test/unit/server/adapters/lib/redis/adapter-cache-redis.test.js
@@ -1,7 +1,7 @@
 const assert = require('assert/strict');
 const sinon = require('sinon');
 const logging = require('@tryghost/logging');
-const RedisCache = require('../index');
+const RedisCache = require('../../../../../../core/server/adapters/lib/redis/AdapterCacheRedis');
 
 describe('Adapter Cache Redis', function () {
     beforeEach(function () {
